moved navbar so separate folder
